requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beanie==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.118.0
h11==0.16.0
idna==3.10
lazy-model==0.3.0
motor==3.7.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pymongo==4.15.2
pytesseract==0.3.13
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
